refactor: rename copilot into pro
