+++ app/web_ui/src/lib/ui/kiln_copilot/assistant_pro_required.svelte
